requirements.txt
fastapi
uvicorn
sse-starlette
python-dotenv
openai
fastmcp
mcp
tavily-python
requests
beautifulsoup4
